# storage/api-docs/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Gestão de Ativos",
    "version": "1.0.0",
    "description": "API para integração e gestão de ativos, bem como exportações e webhooks",
    "contact": {
      "name": "Suporte",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    },
    {
      "url": "https://api.example.com",
      "description": "Production Server"
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Fazer login e obter token",
        "tags": ["Autenticação"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["email", "password"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login bem-sucedido, retorna token",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "token": {
                      "type": "string"
                    },
                    "token_type": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Fazer logout e revogar token",
        "tags": ["Autenticação"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Logout bem-sucedido"
          }
        }
      }
    },
    "/api/user": {
      "get": {
        "summary": "Obter dados do usuário autenticado",
        "tags": ["Usuário"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Dados do usuário"
          }
        }
      }
    },
    "/api/webhooks/register": {
      "post": {
        "summary": "Registrar novo webhook",
        "tags": ["Webhooks"],
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["url", "evento"],
                "properties": {
                  "url": {
                    "type": "string",
                    "format": "url",
                    "example": "https://example.com/webhook"
                  },
                  "evento": {
                    "type": "string",
                    "enum": [
                      "bem.criado",
                      "bem.atualizado",
                      "bem.deletado",
                      "manutencao.criada",
                      "reavaliacacao.criada"
                    ],
                    "example": "bem.criado"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Webhook registrado com sucesso"
          }
        }
      }
    },
    "/api/webhooks/list": {
      "get": {
        "summary": "Listar webhooks registrados",
        "tags": ["Webhooks"],
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Lista de webhooks"
          }
        }
      }
    },
    "/api/webhooks/{id}": {
      "delete": {
        "summary": "Deletar webhook",
        "tags": ["Webhooks"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Webhook deletado com sucesso"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Login com email e password para obter um token"
      }
    }
  }
}